% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_res_split.R
\docType{data}
\name{all_res_split}
\alias{all_res_split}
\title{Parsed and split dataset for all 995 inviduals (498 non-distanced, 497 distanced)}
\format{
A list of 2 lists (\code{$non-distanced} and \code{$distanced}), each themselves
made up of 4 \code{data.frames} or \code{tibbles}:
\describe{
\item{ppt_info}{Complete demographic information plus some simple task-based performance metrics.}
\item{training}{Training data including all missed trials, affect questions, and other metrics.}
\item{test}{Test data.}
\item{gillan_questions}{A \code{data.frame} containing the scores on each of the questions which are
then passed to the classifier to derive transdiagnostic factors. Items not asked are set to 0.}
}
}
\usage{
data(all_res_split)
}
\description{
A dataset containing all output from the online experiment.
}
\keyword{datasets}
