% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_learning_models.R
\name{check_learning_models}
\alias{check_learning_models}
\title{Diagnostic plots and fit metrics for training and test data models}
\usage{
check_learning_models(
  gq,
  mean_pars = TRUE,
  diagnostic_plots = TRUE,
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{gq}{Generated quantities - either a \code{\link[posterior:draws_df]{posterior::draws_df()}} (variational fits) or a vector of
paths to .csv files (either variational or MCMC fits).}

\item{mean_pars}{Output a plot of the mean parameters?}

\item{diagnostic_plots}{Output diagnostic traces and histograms? Requires the \pkg{bayesplot} package.}
}
\description{
Diagnostic plots and fit metrics for training and test data models
}
