% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan_glm.R
\name{cmdstan_glm}
\alias{cmdstan_glm}
\title{Bayesian generalized linear models via CmdStan}
\usage{
cmdstan_glm(
  formula,
  family = gaussian(),
  data,
  weights,
  subset,
  na.action = NULL,
  offset = NULL,
  model = TRUE,
  algorithm = c("sampling", "meanfield", "fullrank"),
  x = FALSE,
  y = TRUE,
  contrasts = NULL,
  out_dir = NULL,
  ...,
  prior = default_prior_coef(family),
  prior_intercept = default_prior_intercept(family),
  prior_aux = exponential(autoscale = TRUE),
  prior_PD = FALSE,
  mean_PPD = !prior_PD,
  sparse = FALSE
)
}
\arguments{
\item{formula, data, subset}{Same as \code{\link[stats]{glm}},
but \emph{we strongly advise against omitting the \code{data}
argument}.}

\item{family}{Same as \code{\link[stats]{glm}}. Accepted families for this
function are \code{gaussian()} or \code{Gamma()}, and any link function.}

\item{na.action, contrasts}{Same as \code{\link[stats]{glm}}, but
rarely specified.}

\item{model, offset, weights}{Same as \code{\link[stats]{glm}}.}

\item{algorithm}{Argument "sampling" is for MCMC (default), while "meanfield"
and "fullrank" are variational algorithms ("meanfield" is the CmdStan
default).}

\item{x}{Logical scalar indicating whether to return the design matrix.}

\item{y}{Logical scalar indicating whether to return the response vector.}

\item{out_dir}{Output directory for model fit environment.}

\item{...}{Further arguments passed to \link[cmdstanr:model-method-sample]{cmdstanr::sample} (i.e.,
\code{refresh}, \code{iter_warmup}, \code{iter_sampling}, \code{chains} etc.)}

\item{prior}{The prior distribution for the (non-hierarchical) regression
coefficients.

The default priors are described in the vignette
\href{http://mc-stan.org/rstanarm/articles/priors.html}{\emph{Prior
Distributions for rstanarm Models}}.
If not using the default, \code{prior} should be a call to one of the
various functions provided by \pkg{rstanarm} for specifying priors. The
subset of these functions that can be used for the prior on the
coefficients can be grouped into several "families":

\tabular{ll}{
\strong{Family} \tab \strong{Functions} \cr
\emph{Student t family} \tab \code{normal}, \code{student_t}, \code{cauchy} \cr
\emph{Hierarchical shrinkage family} \tab \code{hs}, \code{hs_plus} \cr
\emph{Laplace family} \tab \code{laplace}, \code{lasso} \cr
\emph{Product normal family} \tab \code{product_normal} \cr
}

See [http://mc-stan.org/rstanarm/reference/priors.html](here) for details on the
families and how to specify the arguments for all of the functions in the table above.
To omit a prior ---i.e., to use a flat (improper) uniform prior---
\code{prior} can be set to \code{NULL}, although this is rarely a good
idea.

\strong{Note:} Unless \code{QR=TRUE}, if \code{prior} is from the Student t
family or Laplace family, and if the \code{autoscale} argument to the
function used to specify the prior is left at its default and recommended
value of \code{TRUE}, then the default or user-specified prior scale(s) may
be adjusted internally based on the scales of the predictors.}

\item{prior_intercept}{The prior distribution for the intercept (after
centering all predictors, see note below).

The default prior is described in the vignette
\href{http://mc-stan.org/rstanarm/articles/priors.html}{\emph{Prior
Distributions for rstanarm Models}}.
If not using the default, \code{prior_intercept} can be a call to
\code{normal}, \code{student_t} or \code{cauchy}. To omit a
prior on the intercept ---i.e., to use a flat (improper) uniform prior---
\code{prior_intercept} can be set to \code{NULL}.

\strong{Note:} If using a dense representation of the design matrix
---i.e., if the \code{sparse} argument is left at its default value of
\code{FALSE}--- then the prior distribution for the intercept is set so it
applies to the value \emph{when all predictors are centered} (you don't
need to manually center them). This is explained further in the vignette
\href{http://mc-stan.org/rstanarm/articles/priors.html}{\emph{Prior
Distributions for rstanarm Models}}.
If you prefer to specify a prior on the intercept without the predictors
being auto-centered, then you have to omit the intercept from the
\code{\link[stats]{formula}} and include a column of ones as a predictor,
in which case some element of \code{prior} specifies the prior on it,
rather than \code{prior_intercept}. Regardless of how
\code{prior_intercept} is specified, the reported \emph{estimates} of the
intercept always correspond to a parameterization without centered
predictors (i.e., same as in \code{glm}).}

\item{prior_aux}{The prior distribution for the "auxiliary" parameter (if
applicable). The "auxiliary" parameter refers to a different parameter
depending on the \code{family}. For Gaussian models \code{prior_aux}
controls \code{"sigma"}, the error
standard deviation. For negative binomial models \code{prior_aux} controls
\code{"reciprocal_dispersion"}, which is similar to the
\code{"size"} parameter of \code{\link[stats:NegBinomial]{rnbinom}}:
smaller values of \code{"reciprocal_dispersion"} correspond to
greater dispersion. For gamma models \code{prior_aux} sets the prior on
to the \code{"shape"} parameter (see e.g.,
\code{\link[stats:GammaDist]{rgamma}}), and for inverse-Gaussian models it
is the so-called \code{"lambda"} parameter (which is essentially the
reciprocal of a scale parameter). Binomial and Poisson models do not have
auxiliary parameters.

The default prior is described in the vignette
\href{http://mc-stan.org/rstanarm/articles/priors.html}{\emph{Prior
Distributions for rstanarm Models}}.
If not using the default, \code{prior_aux} can be a call to
\code{exponential} to use an exponential distribution, or \code{normal},
\code{student_t} or \code{cauchy}, which results in a half-normal, half-t,
or half-Cauchy prior. See
\href{http://mc-stan.org/rstanarm/reference/priors.html}{\emph{here}}
for details on these functions. To omit a prior ---i.e., to use
a flat (improper) uniform prior--- set \code{prior_aux} to \code{NULL}.}

\item{prior_PD}{A logical scalar (defaulting to \code{FALSE}) indicating
whether to draw from the prior predictive distribution instead of
conditioning on the outcome.}

\item{mean_PPD}{A logical value indicating whether the sample mean of the
posterior predictive distribution of the outcome should be calculated in
the \code{generated quantities} block. If \code{TRUE} then \code{mean_PPD}
is computed and displayed as a diagnostic in the printed output. A useful
heuristic is to check if \code{mean_PPD} is plausible when compared to
\code{mean(y)}. If it is plausible then this does \emph{not} mean that the
model is good in general (only that it can reproduce the sample mean), but
if \code{mean_PPD} is implausible then there may be something wrong, e.g.,
severe model misspecification, problems with the data and/or priors,
computational issues, etc.}

\item{sparse}{A logical scalar (defaulting to \code{FALSE}) indicating
whether to use a sparse representation of the design (X) matrix.
If \code{TRUE}, the the design matrix is not centered (since that would
destroy the sparsity) and likewise it is not possible to specify both
\code{QR = TRUE} and \code{sparse = TRUE}. Depending on how many zeros
there are in the design matrix, setting \code{sparse = TRUE} may make
the code run faster and can consume much less RAM.}
}
\value{
A \code{cmdstanr::CmdStanMCMC()} object.
}
\description{
Generalized linear modeling for Gaussian and gamma responses, with optional
prior distributions for the coefficients, intercept, and auxiliary
parameters.
}
\details{
The \code{stan_glm} function is similar in syntax to
\code{\link[stats]{glm}} but rather than performing maximum likelihood
estimation of generalized linear models, full Bayesian estimation is
performed (if \code{algorithm} is \code{"sampling"}) via MCMC. The Bayesian
model adds priors (independent by default) on the coefficients of the GLM.
}
