% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_subsample.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example probabilistic selection task data}
\format{
A list with four elements (\code{ppt_info}, \code{training},
\code{test}, \code{gillan_questions}).
}
\usage{
data(example_data)
}
\description{
An example dataset with data from ten individuals (non-distanced).
}
\keyword{example_data}
