% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_posterior_quantities.R
\name{generate_posterior_quantities}
\alias{generate_posterior_quantities}
\title{Generate posterior quantities following MCMC}
\usage{
generate_posterior_quantities(
  fit_mcmc,
  data_list,
  out_dir = "outputs/cmdstan",
  save_model_as = "",
  return_type = "",
  par_chains = getOption("mc.cores", 4)
)
}
\arguments{
\item{fit_mcmc}{CmdStanR model environment or a full path to a saved CmdStanR fit (as an .RDS file).}

\item{data_list}{Raw data that \code{fit} was fit to; either an R object or full path to an .RDS file.}

\item{out_dir}{Path to output directory (relative to working directory).}

\item{save_model_as}{Name to give model; defaults to model name from the fit metadata.}

\item{return_type}{Return something other than the fit environment? Options are "paths" (for .csv file paths), or
"draws_list" which returns only the posterior predictions as a \link[posterior:draws_list]{posterior::draws_list}.}

\item{par_chains}{Maximum number of chains to compute in parallel; defaults to \code{options(mc.cores)} if
this has been set, or 4 if not.}
}
\value{
Either file paths or a \link[posterior:draws_list]{posterior::draws_list}.
}
\description{
\code{generate_posterior_quantities} automates the CmdStanR \code{$generate_quantities} method,
which enables the "generated quantities" block of a Stan model to be sampled from after a model is fit.
This vastly reduces memory requirements, particularly if per-sample posterior predictions are desired.
}
