% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{get_affect_ppc}
\alias{get_affect_ppc}
\title{Extract posterior predictions from affect data models and assess fit}
\usage{
get_affect_ppc(draws, raw, adj)
}
\arguments{
\item{draws}{A \code{\link[posterior:draws_df]{posterior::draws_df()}}. Draws outputted from
\link{fit_learning_model} (as a \code{\link[posterior:draws_list]{posterior::draws_list()}}) should be converted via
\code{\link[posterior:draws_df]{posterior::as_draws_df()}} - this is memory intensive, hence it is not done
internally.}

\item{raw}{Raw data, e.g., saved by \code{\link[=fit_learning_model]{fit_learning_model()}}.}

\item{adj}{Name of the affect adjective - one of "happy", "confident" or
"engaged".}
}
\value{
List containing a dataframe with participant identifiers, numbers,
and \eqn{R^2}, MAE and RMSE for each individual; and a named list (by ID) of
data frames with individuals' mean posterior predictions and raw affect data.
}
\description{
\code{get_affect_ppc} combines posterior predictions contained in a
\code{\link[posterior:draws_df]{posterior::draws_df()}} outputted from a fit model with raw affect ratings,
and returns various fit metrics (\eqn{R^2}, MAE, RMSE), for each individual.
}
\examples{
\dontrun{
fit_affect <- fit_learning_model(
  example_data$nd,
  model = "2a",
  affect = TRUE,
  exp_part = "training"
 )

 fit_ls_happy <- get_affect_ppc(
   draws = fit_affect$draws,
   raw = fit_affect$raw_df,
   adj = "happy"
 )}

}
