% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{get_affect_wts}
\alias{get_affect_wts}
\title{Extract individual-level weight parameters from affect data models}
\usage{
get_affect_wts(summary, adj_order = c("happy", "confident", "engaged"))
}
\arguments{
\item{summary}{A \code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}}.}

\item{adj_order}{Same as \code{\link[=fit_learning_model]{fit_learning_model()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with weights by individual and adjective.
}
\description{
\code{get_affect_wts} obtains weight parameters from affect data models, and
names them with the appropriate affect adjective.
}
\examples{
\dontrun{
fit_affect <- fit_learning_model(
  example_data$nd,
  model = "2a",
  affect = TRUE,
  exp_part = "training"
)

aff_wts <- get_affect_wts(fit_affect$summary)
}

}
