% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds_by_chain.R
\name{get_preds_by_chain}
\alias{get_preds_by_chain}
\title{Get and store posterior predictions for training data}
\usage{
get_preds_by_chain(
  out_files,
  out_dir = "",
  obs_df,
  n_draws_chain,
  save_dir = out_dir,
  test = FALSE,
  prefix = "",
  splits = list(blocks = 1:6, sum_blks = list(c(1, 3), c(4, 6))),
  exclude = NULL,
  memory_save = TRUE,
  ...
)
}
\arguments{
\item{out_files}{Vector of .csv file names which contain posterior predictions (e.g., outputted from
\code{\link[=generate_posterior_quantities]{generate_posterior_quantities()}}).}

\item{out_dir}{Path to output directory (defaults to current working directory).}

\item{obs_df}{Raw training data, e.g., outputted from \code{\link[=fit_learning_model]{fit_learning_model()}} (this is best
as it ensures individuals are matched with the correct predictions.)}

\item{n_draws_chain}{Number of MCMC sampling iterations per chain.}

\item{save_dir}{Directory to save items to, will be created if it does not exist. Defaults to the
directory of the output files.}

\item{prefix}{Optional prefix to add to the saved objects.}

\item{memory_save}{An alternative method to obtain predictions, which loads the predictions for each
individual (across all chains) one-by-one, as opposed to importing all the draws for all individuals.
This will be significantly slower but enables the function to run with very limited RAM.}

\item{...}{Other arguments which are unlikely to be necessary to change: \code{n_trials} (default = 360);
\code{vars} (default = "y_pred"), and \code{pred_types} (default = \code{c("AB", "CD", "EF")}).}
}
\value{
An updated \code{tibble} with summed choices per chain and their overall proportion.
}
\description{
\code{get_preds_by_chain} is a helper function which aims to automate the loading of posterior predictions,
with an optional method to help avoid memory overload (and crashes).
}
