% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_multiple.R
\name{import_multiple}
\alias{import_multiple}
\title{JATOS text file parsing function (multiple results)}
\usage{
import_multiple(
  jatos_txt_file,
  separate = TRUE,
  exclusion = TRUE,
  indiv = FALSE,
  ret_incomplete = FALSE,
  ...
)
}
\arguments{
\item{jatos_txt_file}{JATOS .txt file path.}

\item{separate}{Returns separate lists for non-distanced and distanced
groups?}

\item{exclusion}{Exclude those meeting exclusion criteria from questionnaire
/ hBayesDM outputs?}

\item{indiv}{Return individual-level data? This can then be passed to
\code{\link[=plot_import]{plot_import()}}.
and plotted for one individual at a time.}

\item{ret_incomplete}{Return incomplete datasets? They may require manual
parsing.}

\item{...}{Other arguments passed to \code{\link[=import_single]{import_single()}}.}
}
\value{
\code{list} of \code{tibbles} or a single \code{tibble} if
\code{combine == TRUE}.
}
\description{
\code{import_multiple()} parses a JATOS .txt file containing multiple
participant's results to a (list of) R object(s).
}
\examples{
\dontrun{
# Data not included in package (as ~900MB)
all_res_split <- import_multiple(
  "data-raw/all_995_complete.txt", add_sex = TRUE
)
}

}
