% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_single.R
\name{import_single}
\alias{import_single}
\title{JATOS text file parsing function (single result)}
\usage{
import_single(
  jatos_txt_file,
  accuracy = FALSE,
  task_excl = TRUE,
  hbayesDM = FALSE,
  qstns_gillan = TRUE,
  prolific = TRUE,
  combine = FALSE,
  issues = FALSE,
  incomplete = FALSE,
  add_sex = FALSE,
  prolific_export = "data/prolific_export_complete.csv",
  plot = FALSE,
  pal = NULL,
  font = NULL,
  font_size = 14,
  plot_type = c("tr20", "tr60", "tr_all", "tr_questions", "happy", "confident",
    "engaged", "test_perf"),
  ...
)
}
\arguments{
\item{jatos_txt_file}{JATOS text file path}

\item{accuracy}{Include final block AB accuracy < 0.6 as part of exclusion criteria?}

\item{task_excl}{Use digit span of 0 or > 95\% preference for one button as exclusion criteria?}

\item{hbayesDM}{Return additional df with training data formatted for hBayesDM?}

\item{qstns_gillan}{Return formatted subset of questionnaire questions?}

\item{prolific}{Use prolific IDs as subjID?}

\item{combine}{Combine everything into a single data.frame?}

\item{issues}{Return feedback from end of experiment?}

\item{incomplete}{Is the text file of interest incomplete?}

\item{add_sex}{Add sex to participant info? Imputes (assumed) birth-assigned sex for non-binary
individuals from Prolific export - workaround to enable it to be more reasonably controlled for
in models (given very low numbers of non-binary individuals).}

\item{prolific_export}{Path of prolific export file - only required if \code{add_sex} = \code{TRUE}.}

\item{plot}{Return plot(s) of training/test performance?}

\item{pal}{Define a custom colour palette for the plots - need at least 4 values.}

\item{font}{Use a custom font for the plots? Will likely require \code{extrafont::font_import()} to
be run first.}

\item{font_size}{Base plot font size passed to \code{ggplot2::theme_gray}.}

\item{plot_type}{What plots should be outputted?}

\item{...}{Other arguments (used by \code{pstpipeline::import_multiple}).}
}
\value{
\code{list} of \code{tbl_df}, or a single \code{tbl_df} if combine = TRUE
}
\description{
\code{import_single()} parses a single JATOS .txt file to a (list of) R object(s)
}
