% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_posterior_quantities.R
\name{load_posterior_quantities}
\alias{load_posterior_quantities}
\title{Load posterior predictions for training data}
\usage{
load_posterior_quantities(
  out_files,
  obs_df,
  n_draws_chain,
  out_dir = "",
  n_indiv = NULL,
  prefix = "",
  n_trials = 360,
  sub_dir = "",
  vars = "y_pred",
  split_indiv = FALSE,
  obs = NULL,
  pred_type = c("AB", "CD", "EF"),
  n_chains = NULL,
  chains_saved = FALSE,
  ...
)
}
\arguments{
\item{out_files}{csvs}

\item{out_dir}{path}
}
\value{
Series of .RDS files with averaged predictions at different levels (saved to disk).
}
\description{
\code{load_posterior_quantities} is a helper function which aims to automate the loading of
posterior predictions without crashing R due to memory overload.
}
