% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_glm.R
\name{parameter_glm}
\alias{parameter_glm}
\title{Quantify and plot associations between learning parameters and variables of interest}
\usage{
parameter_glm(
  summary_df = list(),
  raw_df = list(),
  var_of_interest,
  covariates,
  interaction = NULL,
  recode_na = NULL,
  factor_scores = NULL,
  ...
)
}
\arguments{
\item{summary_df}{List of \code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} outputs for the fit(s) of interest.}

\item{raw_df}{List of raw data inputs to the above fits (in the same order). Used to
correctly link subject IDs to independent variables.}

\item{...}{Other arguments to pass to \code{\link[=cmdstan_glm]{cmdstan_glm()}} (e.g., to control
number of warm-up and sampling iterations). In addition, use \code{cores} to chnage the
number of parallel chains to sample from.}

\item{indep_vars}{Vector of independent variables to include in the GLMs.}
}
\value{
A \code{\link[posterior:draws_df]{posterior::draws_df()}}.
}
\description{
\code{parameter_glm} is a wrapper around \code{\link[=cmdstan_glm]{cmdstan_glm()}}, taking as inputs
summary tables and raw data from \code{\link[=fit_learning_model]{fit_learning_model()}}, and outputting
the results of GLMs quantifying the association between the individual-level posterior
means of each parameter and the independent variable(s) of interest. Gamma GLMs with
log link functions are used for learning rate models, while standard Gaussian models with
identity link are used for models with inverse temperature as the response variable.
}
