% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_affect.R
\name{plot_affect}
\alias{plot_affect}
\title{Plot an individual's mean posterior predictions compared to their raw affect
ratings}
\usage{
plot_affect(
  fit_list,
  plt_type,
  adj_order = c("happy", "confident", "engaged"),
  nouns = c("Happiness", "Confidence", "Engagement"),
  id_no = NULL,
  r2_coords = c(0.9, 0.8),
  legend_pos = "right",
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{fit_list}{List of outputs from \link{get_affect_ppc}.}

\item{adj_order}{Same as \code{\link[=fit_learning_model]{fit_learning_model()}}.}

\item{nouns}{Formatted noun versions of the adjectives, in order.}

\item{id_no}{If \code{type == "individual"}, a participant number to plot.
If left as \code{NULL}, defaults to the individual with the median \eqn{R^2}
for each adjective.}

\item{r2_coords}{If \code{type == "individual"}, coordinates to print the
\eqn{R^2} value.}

\item{legend_pos, pal, font, font_size}{Same as \link{plot_import}.}

\item{type}{Either \code{"individual"} or \code{"grouped"}.}
}
\value{
A \code{list} of or single \code{ggplot} object(s) (if
\code{type == "grouped"} and \code{length(adj_order) == 1}).
}
\description{
\code{plot_affect} is capable of plotting either grouped or
individual-level posterior predictions (vs. raw observations) for a defined
list of posterior predictions and/or grouping.
}
