% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\title{Plot raw experiment data (single result)}
\usage{
plot_factors(
  df,
  plot_type,
  colnames,
  titles,
  pal = NULL,
  group = NULL,
  font_size = 11,
  font = NULL
)
}
\arguments{
\item{df}{Data frame with factor scores or questions to plot.}

\item{plot_type}{Plot(s) to output: "factor_hist" (which can be grouped), .}

\item{colnames}{Column names to get data from.}

\item{titles}{Title(s) for the plot(s).}

\item{pal}{Custom colour palette to use.}

\item{group}{Grouping for each of the plots (if relevant).}

\item{font_size}{Base plot font size.}

\item{font}{Use a custom font for the plots?}
}
\value{
\code{list} of \code{ggplot} objects
}
\description{
\code{plot_factors} plots various aspects of the transdiagnostic factor derivation, including
histograms and heatmaps of the questions themselves.
}
