% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_import.R
\name{plot_import}
\alias{plot_import}
\title{Plot raw experiment data}
\usage{
plot_import(
  parsed_df,
  import_single = FALSE,
  id = NULL,
  pal = NULL,
  font = NULL,
  font_size = 14,
  plot_type = c("tr20", "tr60", "tr_all", "tr_questions", "happy", "confident",
    "engaged", "test_perf")
)
}
\arguments{
\item{parsed_df}{\code{pstpipeline::import_single()} or \code{pstpipeline::import_multiple()} output.}

\item{import_single}{Is the output from \code{pstpipeline::import_single()}?}

\item{id}{Prolific or other participant ID to select if only a single partipant's data is desired.}

\item{pal}{Define a custom colour palette for the plots? Otherwise reverts to defaults.}

\item{font}{Use a custom font for the plots? Will likely require \code{extrafont::font_import()} to
be run first.}

\item{font_size}{Base plot font size passed to \code{ggplot2::theme_gray}.}

\item{plot_type}{Select plots as required.}
}
\value{
\code{list} of \code{ggplot} objects
}
\description{
\code{plot_import} plots a single participant's data from \code{pstpipeline::import_single()}, or
a single participant (if \code{!is.null(id)}) or all participants' data from
\code{pstpipeline::import_multiple()}.
}
