% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recovery.R
\name{plot_recovery}
\alias{plot_recovery}
\title{Plots to check parameter recovery}
\usage{
plot_recovery(
  raw_pars,
  sim_pars,
  test = FALSE,
  alpha_par_nms = NULL,
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{raw_pars}{Data frame with factor scores or questions to plot.}

\item{sim_pars}{\code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} containing the fitted parameter values
for the relevant model.}

\item{test}{Boolean indicating whether recovered parameters are from the test
phase.}

\item{alpha_par_nms}{Option to rename learning rate parameters, defaults to
the names from \code{par_df}.}

\item{pal, font, font_size}{Same as \code{\link[=plot_import]{plot_import()}}.}
}
\value{
A named \code{list} of \code{ggplot} objects.
}
\description{
\code{plot_recovery} produces correlation plots between input and output
parameter values, plus a confusability matrix.
}
