% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single.R
\name{plot_single}
\alias{plot_single}
\title{Plot raw experiment data (single result)}
\usage{
plot_single(
  subjID,
  parsed_multiple,
  plot_type = c("tr20", "tr60", "tr_all", "tr_questions", "happy", "confident",
    "engaged", "test_perf")
)
}
\arguments{
\item{subjID}{Prolific or other participant ID to select.}

\item{parsed_multiple}{\code{pstpipeline::import_multiple()} output}

\item{plot_type}{Select plots as required.}
}
\value{
\code{list} of \code{ggplot} objects
}
\description{
\code{plot_single} plots a single participant's data from \code{pstpipeline::import_multiple()}
output.
}
