% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{quantile_hdi}
\alias{quantile_hdi}
\title{Compute quantiles of a probability distrbution based on highest density
intervals (HDIs)}
\usage{
quantile_hdi(var, quantile, transform = FALSE)
}
\arguments{
\item{var}{A vector of representative values from a probability distribution
(e.g., MCMC samples).}

\item{quantile}{A vector of quantiles to return.}

\item{transform}{Are values log-scaled (e.g., estimated from Gamma GLMs)?}
}
\value{
A sorted \code{\link[base:vector]{base::vector()}} with all HDIs plus the min/max/median if
specified.
}
\description{
\code{quantile_hdi} computes any number of highest density quantiles from a
sample of representative values, estimated as shortest credible intervals.
}
