% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_preds_by_chain.R
\name{save_preds_by_chain}
\alias{save_preds_by_chain}
\title{Load posterior predictions for training data}
\usage{
save_preds_by_chain(
  out_files,
  obs_df,
  n_draws_chain,
  no_splits,
  out_dir = "",
  n_indiv = NULL,
  prefix = "",
  n_trials = 360,
  n_trials_type = n_trials/3,
  sub_dir = "",
  vars = "y_pred",
  obs = NULL,
  pred_types = c("AB", "CD", "EF"),
  n_chains = NULL,
  chains_saved = FALSE,
  splits = 1,
  ...
)
}
\arguments{
\item{out_files}{csvs}

\item{out_dir}{path}
}
\value{
Series of .RDS files with averaged predictions at different levels (saved to disk).
}
\description{
\code{load_posterior_quantities} is a helper function which aims to automate the loading of
posterior predictions without crashing R due to memory overload.
}
