% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_QL.R
\name{simulate_QL}
\alias{simulate_QL}
\title{Simulate data from single and dual learning rate Q-learning models}
\usage{
simulate_QL(
  summary_df = NULL,
  sample_size = NULL,
  model = "2a",
  test = FALSE,
  affect = FALSE,
  affect_sfx = NULL,
  prev_sample = NULL,
  raw_df = NULL,
  ...
)
}
\arguments{
\item{summary_df}{\code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} output containing posterior mean
parameter values for a group of individuals. If left as NULL, a random sample
of size \code{sample_size} will be taken from specified distributions.}

\item{sample_size}{How may sets of parameters to sample; defaults to 100 or
the number of individuals in
the \code{summary_df}.}

\item{model}{Learning model to simulate, one of \code{"1a"}, \code{"2a"},
\code{"1a2r"}, or \code{"1a1c"}. Defaults to \code{"2a"}.}

\item{test}{Simulate test choices in addition to training choices?}

\item{affect}{Simulate subjective affect ratings (uses full passage-of-time
model).}

\item{affect_sfx}{Affect model suffix to simulate. Mirrors
\code{\link[=fit_learning_model]{fit_learning_model()}} and can be one of \code{"3wt"}, \code{"4wt_trial"},
\code{"4wt_block"}, \code{"4wt_time"}, \code{"5wt_time"},
\code{"delta"}, or \code{"delta-signed"}. If left as \code{NULL}, defaults
to \code{"4wt_time"}.}

\item{prev_sample}{An optional previous sample of id numbers (if you wish to
simulate data for the same subset of individual parameters across a number
of models).}

\item{raw_df}{Provide the raw data used to fit the data originally, so that
subject IDs can be labelled appropriately.}

\item{...}{Other arguments which can be used to control the parameters of the
Beta/Gaussian distributions from which parameter values are sampled.}
}
\value{
Simulated training data (and test data relevant) for a random or
previously fitted sample of parameter values.
}
\description{
\code{simulate_QL} is a function to simulate data from 1-alpha and 2-alpha
Q-learning models, with an experiment structure identical to that run online.
The parameter values can be from (a sample of) those fitted previously to the
real data, or can be randomly sampled.
}
\examples{
train_sim_2a <- simulate_QL(
  sample_size = 5,
  alpha_pos_dens = c(shape = 2, scale = 0.1), # default
  alpha_neg_dens = c(shape = 2, scale = 0.1), # default
  beta_dens = c(mean = 3, sd = 1) # default
)

}
