% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_subsample.R
\name{take_subsample}
\alias{take_subsample}
\title{Helper function to take a subsample of our parsed list (for demonstration
purposes)}
\usage{
take_subsample(parsed_list, n_ppts)
}
\arguments{
\item{parsed_list}{A list outputted from \code{\link[=import_multiple]{import_multiple()}}.}

\item{n_ppts}{Sample size to take.}
}
\value{
A named \code{list}.
}
\description{
\code{take_subsample} takes a sample of size n to enable model fitting on a
smaller sample.
}
