% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_learning_models.R
\name{check_learning_models}
\alias{check_learning_models}
\title{Diagnostic plots and fit metrics for training and test data models}
\usage{
check_learning_models(
  draws,
  test = FALSE,
  mean_pars = TRUE,
  diagnostic_plots = TRUE,
  alpha_par_nms = NA,
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{draws}{Post-warmup draws - either a \code{\link[posterior:draws_array]{posterior::draws_array()}}, a
\code{\link[posterior:draws_list]{posterior::draws_list()}}, or a vector of file paths to the .csv output
files. May also be a \code{\link[posterior:draws_df]{posterior::draws_df()}} but chain-by-chain diagnostics
will not be possible.}

\item{test}{Boolean indicating whether recovered parameters are from the test
phase.}

\item{mean_pars}{Output a plot of the mean parameters?}

\item{diagnostic_plots}{Output diagnostic traces and histograms? Requires the
\pkg{bayesplot} package.}

\item{alpha_par_nms}{Option to rename learning rate parameters for models
with more than one.
the names from \code{draws}.}

\item{pal, font, font_size}{Same as \link{plot_import}.}
}
\description{
This function is called automatically by \code{\link[=fit_learning_model]{fit_learning_model()}} when
\code{model_checks = TRUE}, but can also be run separately if desired.
}
