% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example probabilistic selection task data}
\format{
A list with two elements: \code{nd} has ten non-distanced
individuals, and \code{dis} has ten distanced individuals. Each sub-list is a
list with four elements: (\code{ppt_info}, \code{training}, \code{test}, and
\code{gillan_questions}).
}
\usage{
data(example_data)
}
\description{
An example dataset with data from ten individuals from each group.
}
\keyword{example_data}
