% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_single.R
\name{import_single}
\alias{import_single}
\title{JATOS text file parsing function (single result)}
\usage{
import_single(
  jatos_txt_file,
  accuracy = FALSE,
  task_excl = TRUE,
  hbayesDM = TRUE,
  qstns_gillan = TRUE,
  prolific = TRUE,
  combine = FALSE,
  issues = FALSE,
  incomplete = FALSE,
  plot = FALSE,
  plot_type = c("tr20", "tr60", "tr_all", "tr_questions", "happy", "confident",
    "engaged", "test_perf"),
  ...
)
}
\arguments{
\item{jatos_txt_file}{JATOS text file path}

\item{accuracy}{Include final block AB accuracy < 0.6 as part of exclusion criteria?}

\item{task_excl}{Use digit span == 0 or > 95\% preference for one button as exclusion criteria?}

\item{hbayesDM}{Return additional df with training data formatted for hBayesDM?}

\item{qstns_gillan}{Return formatted subset of questionnaire questions?}

\item{prolific}{Use prolific IDs as subjID?}

\item{combine}{Combine everything into a single data.frame?}

\item{issues}{Return feedback from end of experiment?}

\item{incomplete}{Is the text file of interest incomplete?}

\item{plot}{Return plot(s) of training/test performance?}

\item{plot_type}{What plots should be outputted?}

\item{...}{Other arguments (used by \code{pstpipeline::import_multiple}).}
}
\value{
\code{list} of \code{tbl_df}, or a single \code{tbl_df} if combine = TRUE
}
\description{
\code{import_single()} parses a single JATOS .txt file to a (list of) R object(s)
}
