% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_glm.R
\name{parameter_glm}
\alias{parameter_glm}
\title{Quantify and plot associations between learning parameters and variables of
interest}
\usage{
parameter_glm(
  summary_df = list(),
  raw_df = list(),
  var_of_interest,
  covariates,
  affect_number = NULL,
  interaction = NULL,
  recode_na = NULL,
  factor_scores = NULL,
  rhat_upper = 1.1,
  ess_lower = 100,
  ...
)
}
\arguments{
\item{summary_df}{List of \code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} outputs for the fit(s) of
interest.}

\item{raw_df}{List of raw data inputs to the above fits (in the same order).
Used to correctly link subject IDs to independent variables.}

\item{var_of_interest}{Variable of interest.}

\item{covariates}{Vector of covariates to control for in the GLMs.}

\item{affect_number}{For affect model fits, specify the number (i.e., 1, 2,
or 3) of the affect noun/verb of interest. If affect parameters are found in
model summaries, and this is not specified, GLMs will default to standard
Q-learning parameters.}

\item{interaction}{Optional variable to interact with the variable of
interest. The GLMs will then be run twice with this variable reverse coded
the second time to obtain posterior samples for the variable of interest in
both groups. This variable must be binary and only 1 interaction is allowed.}

\item{recode_na}{Some demographic questions were conditional, and so there
exist NAs. This argument allows these terms to be recoded as appropriate
(in all binary cases, this should be set to 0).}

\item{factor_scores}{Given the factor scores were derived separately, this
argument allows the \code{data.frame} containing the factor scores to be
supplied, so that these factors can be included in models.}

\item{rhat_upper, ess_lower}{Same as \code{\link[=plot_raincloud]{plot_raincloud()}}.}

\item{...}{Other arguments to pass to \code{\link[=cmdstan_glm]{cmdstan_glm()}} (e.g., to control
number of warm-up and sampling iterations). In addition, use \code{cores} to
change the number of parallel chains to sample from.}
}
\value{
A \code{\link[posterior:draws_df]{posterior::draws_df()}}.
}
\description{
\code{parameter_glm} is a wrapper around \code{\link[=cmdstan_glm]{cmdstan_glm()}}, taking as inputs
summary tables and raw data from \code{\link[=fit_learning_model]{fit_learning_model()}}, and outputting
the results of GLMs quantifying the association between the individual-level
posterior means of each parameter and the independent variable(s) of
interest. Gamma GLMs with log link functions are used for learning rate and
decay factor models (i.e., positively skewed and constrained between 0 and 1)
while standard Gaussian models with identity link functions are used for
models with inverse temperature or weights as the response variable.
}
