% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\title{Various plots for factor prediction and derivation}
\usage{
plot_factors(
  df,
  plot_type,
  colnames = NA,
  titles = NA,
  r2 = NA,
  qn = NA,
  hyp_alph = 0.1,
  grouped = FALSE,
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{df}{Data frame with factor scores or questions to plot. For
\code{predictive} and \code{factor_htmp} plots, this should be a named list
of \code{data.frames} (including "preds" & "scores" or "qns" & "coefs"
respectively). If a \code{factor_hist} plot with \code{grouped = TRUE} is
desired, it should be a list of \code{data.frames} (one per group).}

\item{plot_type}{Plot(s) to output: \code{factor_hist} (which can be
grouped), \code{r2_plot},\code{predictive}, and \code{factor_htmp}.}

\item{colnames}{Column names to get data from.}

\item{titles}{Title(s) for the plot(s) or pretty names more generally.}

\item{r2}{Array of \eqn{R^{2}}{R^2} values.}

\item{qn}{Number of questions - used to find correct \eqn{R^{2}}{R^2} values.}

\item{hyp_alph}{Chosen alpha value (used to draw a dotted line on an
\code{r2_plot}).}

\item{grouped}{.}

\item{pal, font, font_size}{Same as \link{plot_import}.}
}
\value{
A single or \code{list} of \code{ggplot} object(s).
}
\description{
\code{plot_factors} plots various aspects of the transdiagnostic factor
derivation, including histograms and heatmaps of the questions themselves.
}
\examples{
# Please this notebook for usage (examples would require Python input):
# https://github.com/qdercon/pstpipeline/blob/main/notebooks/data_cleaning_factor_derivation.ipynb # nolint
}
