% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_glm.R
\name{plot_glm}
\alias{plot_glm}
\title{Plot associations between Q-learning model parameters and variables of interest}
\usage{
plot_glm(
  par_df,
  plot_var,
  id.col = "parameter",
  grp = id.col,
  grp_labs = NULL,
  ovrll_title = NULL,
  cred = c(0.95, 0.99),
  top_right = TRUE,
  coord_flip = TRUE,
  dist_nudge = 0,
  max_dist_width = 0.5,
  point_alpha = 0.25,
  point_size = 0.75,
  box_alpha = 0.6,
  box_width = 0.125,
  box_nudge = 0.1,
  pal = NULL,
  font_size = 11,
  font = ""
)
}
\arguments{
\item{par_df}{A \code{\link[posterior:draws_df]{posterior::draws_df()}}, likely obtained from running
\code{\link[=parameter_glm]{parameter_glm()}}.}

\item{plot_var}{The variable of interest to plot (e.g., distanced vs non-distanced).}

\item{id.col}{The column that contains the QL model parameter names.}

\item{grp}{Optional group to plot separately on each plot, which should be the interaction
variable specified in \code{\link[=parameter_glm]{parameter_glm()}}.}

\item{grp_labs}{Optional labels for the groups defined by \code{grp}. It is recommended to
first run the function with this kept as \code{NULL} to make sure you label the correct
densities.}

\item{ovrll_title}{Title to set for the whole plot.}

\item{cred}{Vector, length 2, which defines the \% HDI covered by the boxplot boxes and
lines respectively.}

\item{top_right, coord_flip}{Booleans; if \code{TRUE} the densities will be on the top
or the right depending on whether \code{coord_flip} is \code{TRUE} or \code{FALSE}
respectively.}

\item{dist_nudge, max_dist_width, point_alpha, point_size}{Control the position and size of
the density, or the transparency and size of the points that make it up respectively.}

\item{box_alpha, box_width, box_nudge}{Control the transparency, size, and position of the
summary boxplot.}

\item{pal, font_size, font}{Same as \code{\link[=plot_import]{plot_import()}}.}
}
\description{
\code{plot_glm} is designed to plot the results of fitting Bayesian GLMs, with
QL model parameters as the response variables, via the \code{\link[=parameter_glm]{parameter_glm()}}
function. It plots both a box-and-whisker plot (defaulting to 95\% HDIs for the box,
and 99\% HDIs for the lines), plus the posterior distribution of coefficients, using
a version of \code{\link[ggbeeswarm:geom_quasirandom]{ggbeeswarm::geom_quasirandom()}} which has been modified to output
half-violin plots made up of (half of) the individual posterior draws.
}
