% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_import.R
\name{plot_import}
\alias{plot_import}
\title{Plot raw experiment data}
\usage{
plot_import(
  parsed_list,
  import_single = FALSE,
  id = NULL,
  types = c("train", "test", "affect"),
  plt.train = list(),
  plt.test = list(),
  plt.affect = list(),
  grp_compare = NULL,
  grp_names = c(),
  recode_na = NULL,
  aff_by_reward = FALSE,
  legend_pos = "right",
  pal = NULL,
  font = "",
  font_size = 14,
  ...
)
}
\arguments{
\item{parsed_list}{\code{\link[=import_single]{import_single()}} or \code{\link[=import_multiple]{import_multiple()}} output.}

\item{import_single}{Is the output from \code{\link[=import_single]{import_single()}}?}

\item{id}{subjID to select if only plots for a single participant are
desired. Will also accept a single numeric value i, which will select the
ith participant in the output.}

\item{types}{Types of plot to output, choose from any (or all) of
\code{train}, \code{test}, and \code{affect}.}

\item{plt.train}{List of length <= 2, with the first element a single value
or numeric vector of the number of trials to lag in the calculation of
cumulative probabilities, and the second element a vector of training types
to include.}

\item{plt.test}{List of length <= 2. The first element specifies the types
of test pairs to plot; accepted inputs include "chooseA", "avoidB", "novel",
"training", "all", and/or character vector(s) of specific pairs. The second
input defines how these options are plotted - either "grouped" or
"individual"; note "grouped" plots will not work for custom inputs or if
"all" is selected as an option.}

\item{plt.affect}{List of length <= 2 indicating (1) how many trials to lag
(only a single value accepted), and (2) the nouns to plot (can be any of
"happy", "confident", "engaged", or "fatigue").}

\item{grp_compare}{Group to compare on which is found from the participant
info. Note that if \code{parsed_list} is split into 2 (i.e., distanced and
non-distanced), comparisons will be automatically made on this split.}

\item{grp_names}{Vector of labels for plot keys for the different groups.
An attempt will be made to label these automatically; it is recommended to
first leave this list empty to make sure the correct labels are applied.}

\item{recode_na}{Some grouping variables are \code{NA} in the participant
information due to them being asked conditionally.  This option enables these
to be recoded as appropriate (e.g., to 0 or \code{FALSE}).}

\item{aff_by_reward}{Enables affect plots to be compared by whether or not
the prior stimulus was rewarded or not. This will override grp_compare (but
won't affect other types of plots).}

\item{legend_pos}{Enables the legend positions to be set manually.}

\item{pal}{Define a custom colour palette for the plots? Otherwise reverts to
defaults.}

\item{font}{Use a custom font for the plots?}

\item{font_size}{Base plot font size.}

\item{...}{Other arguments, used internally by other functions calling this
one.}
}
\value{
Either a single or named \code{list} of \code{ggplot} objects
}
\description{
\code{plot_import} plots a single participant's data from \code{\link[=import_single]{import_single()}},
or a single participant (if \code{!is.null(id)}), or all participants' data
from \code{\link[=import_multiple]{import_multiple()}}.
}
