% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recovery.R
\name{plot_recovery}
\alias{plot_recovery}
\title{Plots to check parameter recovery}
\usage{
plot_recovery(
  raw_pars,
  sim_pars,
  test = FALSE,
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{raw_pars}{Data frame with factor scores or questions to plot.}

\item{sim_pars}{\code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} containing the fitted parameter values for the relevant model.}

\item{pal}{Custom colour palette to use.}

\item{font}{Use a custom font for the plots? Warnings suggest \code{extrafont::font_import()} should
be run.}

\item{font_size}{Base plot font size.}
}
\value{
A named \code{list} of \code{ggplot} objects.
}
\description{
\code{plot_recovery} produces correlation plots between input and output parameter values,
plus a confusability matrix.
}
