% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recovery.R
\name{plot_recovery}
\alias{plot_recovery}
\title{Plots to check parameter recovery}
\usage{
plot_recovery(
  raw_pars,
  rec_pars,
  test = FALSE,
  affect = FALSE,
  alpha_par_nms = NULL,
  plot_together = TRUE,
  incl_legend = !plot_together,
  pal = NULL,
  font = "",
  font_size = 11
)
}
\arguments{
\item{raw_pars}{Parameters outputted from \code{\link[=simulate_QL]{simulate_QL()}}.}

\item{rec_pars}{\code{\link[cmdstanr:fit-method-summary]{cmdstanr::summary()}} containing the fitted parameter values
for the relevant model.}

\item{test}{Boolean indicating whether recovered parameters are from the test
phase.}

\item{affect}{Boolean indicating whether recovered parameters are from an
affect model.}

\item{alpha_par_nms}{Option to rename learning rate parameters, defaults to
the names from \code{par_df}.}

\item{plot_together}{If \code{TRUE}, returns a panel with all plots plotted,
otherwise a named list of plots is returned.}

\item{incl_legend}{Boolean indicating whether to include legends for affect
parameter plots (for the question types).}

\item{pal, font, font_size}{Same as \code{\link[=plot_import]{plot_import()}}.}
}
\value{
A named \code{list} of \code{ggplot} objects.

A list or grid of ggplots.
}
\description{
\code{plot_recovery} produces correlation plots between input and output
parameter values, plus a confusability matrix.
}
\examples{
\dontrun{
dir.create("outputs/cmdstan/simulated_data")

train_sim_2a <- simulate_QL(
  sample_size = 10,
  alpha_pos_dens = c(shape = 2, scale = 0.1), # default
  alpha_neg_dens = c(shape = 2, scale = 0.1), # default
  beta_dens = c(mean = 3, sd = 1) # default
)

mcmc_2a_train_sim <- fit_learning_model(
  train_sim_2a$sim, model = "2a", exp_part = "training",
  vb = FALSE, model_checks = FALSE, par_recovery = TRUE,
  outputs = c("raw_df", "stan_datalist", "summary", "draws_list"),
  out_dir = "outputs/cmdstan/simulated_data"
)

plot_recovery(train_sim_2a$pars, mcmc_2a_train_sim$summary)
}

}
