% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{quantile_hdi}
\alias{quantile_hdi}
\title{Compute quantiles of a probability distrbution based on highest density
intervals (HDIs)}
\usage{
quantile_hdi(var, quantile, ...)
}
\arguments{
\item{var}{A vector of representative values from a probability distribution
(e.g., MCMC samples).}

\item{quantile}{A vector of quantiles to return.}

\item{...}{Internal arguments.}
}
\value{
A sorted vector with all specified quantiles.
}
\description{
\code{quantile_hdi} computes any number of highest density quantiles from a
sample of representative values, estimated as shortest credible intervals. If
\code{quantile} contains 0, 0.5, or 1, will return the minimum, median, and
maximum respectively.
}
\examples{
p_density <- rnorm(100, 2, 0.5)
quantile_hdi(p_density, c(0, 0.025, 0.5, 0.0975, 1))
}
