% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_subsample.R
\name{take_subsample}
\alias{take_subsample}
\title{Helper function to take a subsample of our parsed list (for demonstration purposes)}
\usage{
take_subsample(parsed_list, n_ppts)
}
\arguments{
\item{parsed_list}{An \code{pstpipeline::import_multiple()} output.}

\item{n_ppts}{Sample size to take.}
}
\value{
A named \code{list}.
}
\description{
\code{take_subsample} takes a \code{pstpipeline::import_multiple()} output, and takes a sample of size n.
}
